moves activate deactivate and run in common file for sources
