+++ components/crove_app/sources/document_submitted/document_submitted.mjs
@@ -1,68 +1,18 @@
-import croveApp from "../../crove_app.app.mjs";
-import { v4 as uuid } from "uuid";
+import common from "../common.mjs";
 
 export default {
+  ...common,
   key: "crove_app-document-submitted",
   name: "Document Submitted",
   description: "Triggers when a document is submitted.",
   version: "0.0.1",
   type: "source",
-  props: {
-    croveApp,
-    db: "$.service.db",
-    http: {
-      type: "$.interface.http",
-      customResponse: true,
+  methods: {
+    ...common.methods,
+    getEvents() {
+      return [
+        "document.response.submitted",
+      ];
     },
-    template_id: {
-      propDefinition: [
-        croveApp,
-        "template_id",
-      ],
-    },
-  },
-  hooks: {
-    async activate() {
-      const validationToken = uuid();
-
-      let config = {
-        url: `${this.croveApp._getBaseUrl()}/webhooks/templates/create/`,
-        method: "POST",
-        data: {
-          template: this.template_id,
-          name: `pipedream-webhook: ${validationToken}`,
-          webhook_url: this.http.endpoint,
-          method: "POST",
-          events: [
-            "document.response.submitted",
-          ],
-        },
-      };
-
-      const response = await this.croveApp._makeRequest(config);
-
-      this.db.set("webhookId", response.id);
-    },
-    async deactivate() {
-      const webhookId = this.db.get("webhookId");
-      let config = {
-        url: `${this.croveApp._getBaseUrl()}/webhooks/templates/${webhookId}/`,
-        method: "DELETE",
-      };
-      await this.croveApp._makeRequest(config);
-    },
-  },
-  async run(event) {
-    const { body } = event;
-
-    this.http.respond({
-      status: 200,
-    });
-
-    this.$emit(body, {
-      id: body.webhook.id,
-      summary: `New event ${body.webhook.id} received`,
-      ts: new Date(),
-    });
   },
 };
